cli: move has_tracked_remote_bookmarks() to cli_util
