+++ cli/src/commands/bookmark/rename.rs
@@ -16,7 +16,7 @@ use clap_complete::ArgValueCandidates;
 use jj_lib::op_store::RefTarget;
 use jj_lib::ref_name::RefNameBuf;
 
-use super::has_tracked_remote_bookmarks;
+use crate::cli_util::has_tracked_remote_bookmarks;
 use crate::cli_util::CommandHelper;
 use crate::command_error::user_error;
 use crate::command_error::CommandError;
